(Refactor) : domain package refactoring
(Refactor) : domain package refactoring

warehouse > receiving -> warehouse > domain > receiving
warehouse > users -> warehouse > domain > users
