+++ warehouse/src/main/java/warehouse/domain/receiving/temp.java
@@ -1,4 +1,5 @@
-package warehouse.users;
+package warehouse.domain.receiving;
+
 // TODO 삭제 예정 클래스입니다.
 public class temp {
