+++ warehouse/src/main/java/warehouse/domain/users/temp.java
@@ -1,4 +1,5 @@
-package warehouse.receiving;
+package warehouse.domain.users;
+
 // TODO 삭제 예정 클래스입니다.
 public class temp {
